requirements.txt
fastapi
uvicorn[standard]
httptools
jinja2
sqlalchemy[asyncio]
asyncpg
pymongo
redis
aiobotocore
alembic
python-dotenv
strawberry-graphql[fastapi]
argon2-cffi
boto3
htmx
fasthtml[oauth]
MonsterUI
stripe